multi-agent-load-test/pyproject.toml
[tool.poetry]
name = "multi-load-test"
version = "0.1.0"
description = ""
authors = ["Jason Sherman <tools@usingtechnolo.gy>"]

[tool.poetry.dependencies]
python = "^3.9"
locust = "^2.15.1"
Random-Word = "^1.0.11"
poetry-dotenv-plugin = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

